Combine TabixSam and TabixSamWithPolyA
